finition de la page producteurs
